Moved core Compile functions to ACompiler
